test: inline analysis_core fixture loader into spec to match allowed-files policy
OPE-14's allowed-files list restricts __test__ contributions to *.spec.ts
files. Inline the fixture loader and OHLCV dataset builder into
legacy-parity.spec.ts and drop the standalone helper module so every
file under src/domain/analysis/__test__/ is a *.spec.ts.
